pyproject.toml
[project]
name = "pixelproof"
version = "1.0.0"
description = "Detect fake, AI-generated, and Photoshopped images with forensic-grade analysis"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
keywords = [
    "forensics",
    "exif",
    "photo",
    "fake",
    "deepfake",
    "photoshop",
    "ela",
    "image-analysis",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Security",
]
dependencies = ["Pillow>=10.0.0"]

[project.optional-dependencies]
pdf = ["markdown2>=2.4.0", "weasyprint>=60.0"]

[project.scripts]
pixelproof = "pixelproof:main"
pixelproof-deep = "deep_analysis:main"
pixelproof-pdf = "generate_pdf:main"

[project.urls]
Homepage = "https://github.com/kevinthomas/pixelproof"
Issues = "https://github.com/kevinthomas/pixelproof/issues"

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"
